+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Withdraw notification when message has been read.
+ 
 ## [0.14.1] - 2024-03-22
 
 ### Added